Moved models to API, simplified MediatR config
