+++ src/FocusAPI/Methods/Shop/GetShopData.cs
@@ -2,15 +2,15 @@
 using FocusCore.Queries;
 using MediatR;
 
-namespace FocusCore.Methods.User;
-public class GetUserData
+namespace FocusApi.Methods.Shop;
+public class GetShopData
 {
-    public class Handler : IRequestHandler<GetUserDataQuery, UserModel>
+    public class Handler : IRequestHandler<GetShopDataQuery, UserModel>
     {
         // TODO: Add db client references
         public Handler() { }
 
-        public async Task<UserModel> Handle(GetUserDataQuery query, CancellationToken cancellationToken)
+        public async Task<UserModel> Handle(GetShopDataQuery query, CancellationToken cancellationToken)
         {
             return new UserModel { Id = Guid.NewGuid(), Name = "Gunter" };
         }
