+++ src/FocusAPI/Program.cs
@@ -1,4 +1,3 @@
-using FocusCore.Methods.User;
 using System.Reflection;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -11,7 +10,7 @@ builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 
 // Configure MediatR
-Assembly[] assemblies = new Assembly[] { Assembly.GetExecutingAssembly(), Assembly.GetAssembly(typeof(GetUserData)) };
+Assembly[] assemblies = [Assembly.GetExecutingAssembly()];
 builder.Services.AddMediatR(cfg => cfg.RegisterServicesFromAssemblies(assemblies));
 
 var app = builder.Build();
